pyproject.toml
[project]
name = "les-mosquitos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.17.1",
    "drf-nested-routers>=0.95.0",
    "pillow>=12.1.1",
    "psycopg[binary]",
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
    "requests>=2.33.1",
    "ruff>=0.15.8",
]

[dependency-groups]
prod = [
    "gunicorn>=25.3.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "les_mosquitos.settings"
